delete confirm based on os

+++ www/js/controllers.js
@@ -19,8 +19,8 @@ angular.module('app.controllers', [])
     });	
     /* Delete method */
         $scope.deleteUser = function(userID,index) {
-            
-            /*var confirmDelete = $ionicPopup.confirm({
+        if(isAndroid) {
+            var confirmDelete = $ionicPopup.confirm({
              title: 'Delete Confirmation',
              template: 'Are you sure you want to delete?'
             });
@@ -32,30 +32,27 @@ angular.module('app.controllers', [])
              }else{
                 $ionicListDelegate.closeOptionButtons(); //hide the guesture on prompt cancel
              } 
-           });*/
-            
+           });
+        }else if(isIOS){
             var hideSheet = $ionicActionSheet.show({
-            titleText: 'Are you sure you want to delete?',
-            cancelText: 'Cancel',
-            destructiveText: 'Delete',
-            cancel: function () {
-              // if the user cancel's deletion, hide the list item's delete button
-              $ionicListDelegate.closeOptionButtons();
-            },
-            destructiveButtonClicked: function () {
-              // delete expense by its id property            
-              $http.delete(base_url+"api/v1/deleteuser/"+userID);
-              $scope.users.splice(index, 1); //instantly removes item
-
-              // hide the confirmation dialog
-              hideSheet();
-            }
-          });
-
-            
-            
-            
-            
+                titleText: 'Are you sure you want to delete?',
+                cancelText: 'Cancel',
+                destructiveText: 'Delete',
+                cancel: function () {
+                  // if the user cancel's deletion, hide the list item's delete button
+                  $ionicListDelegate.closeOptionButtons();
+                },
+                destructiveButtonClicked: function () {
+                  // delete expense by its id property            
+                  $http.delete(base_url+"api/v1/deleteuser/"+userID);
+                  $scope.users.splice(index, 1); //instantly removes item
+                  // hide the confirmation dialog
+                  hideSheet();
+                }
+            });
+        }
+        
+   
             
         }
     /* Delete method */
